rename Libra to Diem
Found when trying to visit the old Libra repo: 

Note to readers: On December 1, 2020, the Libra Association was renamed to Diem Association. The project repos are in the process of being migrated. All projects will remain available for use here until the migration to a new GitHub Organization is complete.

Note that IPLD logo isn't displaying properly on this site. IPFS and libp2p logos display for me on the site but not in this GitHub preview.